+++ components/src/Table/withSelectableColumn.js
@@ -1,18 +1,31 @@
 import React from "react";
+import PropTypes from "prop-types";
 import { Checkbox } from "../Checkbox";
+import { rowPropType } from "./Table.types";
 
 const SELECTABLE_COLUMN_DATA_KEY = "selected";
 
-const selectHeaderRenderer = (onSelectHeader, isHeaderSelected) => () => (
+const selectHeaderFormatter = (onSelectHeader, isHeaderSelected) => () => (
   <Checkbox checked={isHeaderSelected} onChange={onSelectHeader} aria-label="toggle all row selections" />
 );
 
-const selectCellRenderer = onSelectRow => (cellData, columnData, row) => {
+const SelectCell = ({ row, onSelectRow }) => {
   const selectRowHandler = () => onSelectRow(row);
   return (
     <Checkbox aria-label="toggle row selection" checked={row[SELECTABLE_COLUMN_DATA_KEY]} onChange={selectRowHandler} />
   );
 };
+
+SelectCell.propTypes = {
+  row: rowPropType.isRequired,
+  onSelectRow: PropTypes.func
+};
+
+SelectCell.defaultProps = {
+  onSelectRow: null
+};
+
+const selectCellRenderer = onSelectRow => props => <SelectCell onSelectRow={onSelectRow} {...props} />;
 
 const addSelectableColumn = ({
   columns,
@@ -26,7 +39,7 @@ const addSelectableColumn = ({
   const selectableColumn = {
     dataKey: SELECTABLE_COLUMN_DATA_KEY,
     cellRenderer: selectCellRenderer(onSelectRow),
-    headerRenderer: selectHeaderRenderer(onSelectHeader, isHeaderSelected)
+    headerFormatter: selectHeaderFormatter(onSelectHeader, isHeaderSelected)
   };
   const selectableCellData = rowKey => ({
     [SELECTABLE_COLUMN_DATA_KEY]: selectedRows.includes(rowKey)