+++ docs/src/components/PropsTable.js
@@ -2,11 +2,15 @@ import React from "react";
 import PropTypes from "prop-types";
 import { Table, Text } from "@nulogy/components";
 
-const smallTextRenderer = cellData => (
+const smallTextRenderer = ({ cellData }) => (
   <Text py="x1" fontSize="small">
     {cellData}
   </Text>
 );
+
+smallTextRenderer.propTypes = {
+  cellData: PropTypes.string.isRequired
+};
 
 const columns = [
   {